requirements.txt
numpy==1.15.4
pillow==5.4.1
tensorflow==1.13.1
opencv-python==4.0.0.21